changed all encoders to getDistance() and changed initCount's to doubles

+++ src/org/usfirst/frc/team3205/robot/commands/autoDriveBackSlow.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj.command.Command;
  */
 public class autoDriveBackSlow extends Command {
 
-	private int initCount;
+	private double initCount;
     public autoDriveBackSlow() {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
